added missing urls from issues & prs
