+++ src/request_interceptors.js
@@ -1,40 +1,5 @@
 import {getTwitterReferer} from './utils';
-
-const urlsList = [
-  'https://medium.com/*',
-  'https://www.google.com/search/*',
-  'https://towardsdatascience.com/*',
-  'https://hackernoon.com/*',
-  'https://medium.freecodecamp.org/*',
-  'https://psiloveyou.xyz/*',
-  'https://betterhumans.coach.me/*',
-  'https://codeburst.io/*',
-  'https://theascent.pub/*',
-  'https://*.medium.com/*',
-  'https://medium.mybridge.co/*',
-  'https://uxdesign.cc/*',
-  'https://levelup.gitconnected.com/*',
-  'https://itnext.io/*',
-  'https://entrepreneurshandbook.co/*',
-  'https://proandroiddev.com/*',
-  'https://blog.prototypr.io/*',
-  'https://thebolditalic.com/*',
-  'https://blog.usejournal.com/*',
-  'https://blog.angularindepth.com/*',
-  'https://blog.bitsrc.io/*',
-  'https://blog.devartis.com/*',
-  'https://blog.maddevs.io/*',
-  'https://blog.getambassador.io/*',
-  'https://uxplanet.org/*',
-  'https://instagram-engineering.com/*',
-  'https://calia.me/*',
-  'https://productcoalition.com/*',
-  'https://engineering.opsgenie.com/*',
-  'https://android.jlelse.eu/*',
-  'https://robinhood.engineering/*',
-  'https://blog.hipolabs.com/*',
-  'https://ux.shopify.com/*',
-];
+import urlsList from './urls';
 
 export default function intercept() {
   function onBeforeSendHeaders(details) {
