+++ src/url.js
@@ -1,0 +1,48 @@
+const urlsList = [
+    'https://medium.com/*',
+    'https://www.google.com/search/*',
+    'https://towardsdatascience.com/*',
+    'https://hackernoon.com/*',
+    'https://medium.freecodecamp.org/*',
+    'https://psiloveyou.xyz/*',
+    'https://betterhumans.coach.me/*',
+    'https://codeburst.io/*',
+    'https://theascent.pub/*',
+    'https://*.medium.com/*',
+    'https://medium.mybridge.co/*',
+    'https://uxdesign.cc/*',
+    'https://levelup.gitconnected.com/*',
+    'https://itnext.io/*',
+    'https://entrepreneurshandbook.co/*',
+    'https://proandroiddev.com/*',
+    'https://blog.prototypr.io/*',
+    'https://thebolditalic.com/*',
+    'https://blog.usejournal.com/*',
+    'https://blog.angularindepth.com/*',
+    'https://blog.bitsrc.io/*',
+    'https://blog.devartis.com/*',
+    'https://blog.maddevs.io/*',
+    'https://blog.getambassador.io/*',
+    'https://uxplanet.org/*',
+    'https://instagram-engineering.com/*',
+    'https://calia.me/*',
+    'https://productcoalition.com/*',
+    'https://engineering.opsgenie.com/*',
+    'https://android.jlelse.eu/*',
+    'https://robinhood.engineering/*',
+    'https://blog.hipolabs.com/*',
+    'https://ux.shopify.com/*',
+    'https://engineering.talkdesk.com/*',
+    "https://blog.codegiant.io",
+    "https://tech.olx.com/*",
+    "https://netflixtechblog.com/*",
+    "https://hackingandslacking.com/*",
+    "https://blog.kotlin-academy.com/*",
+    "https://blog.securityevaluators.com/*",
+    "https://blog.kubernauts.io/*",
+    "https://blog.coffeeapplied.com/*",
+    "https://unbounded.io/*",
+    "https://writingcooperative.com/*",
+];
+
+export default urlsList;
